Refactor LoadGen to pull out tx generation code
