+++ src/simulation/test/LoadGeneratorTests.cpp
@@ -287,8 +287,13 @@ TEST_CASE("generate soroban load", "[loadgen][soroban]")
         rand_uniform<int64_t>(INT64_MAX - 10'000, INT64_MAX);
     upgradeCfg.startingEvictionScanLevel = rand_uniform<uint32_t>(4, 8);
 
-    auto upgradeSetKey =
-        loadGen.getConfigUpgradeSetKey(createUpgradeLoadGenConfig);
+    auto testingKeys =
+        app.getLoadGenerator().getContractInstanceKeysForTesting();
+    REQUIRE(testingKeys.size() == 1);
+    auto contractId = testingKeys.begin()->contractData().contract.contractId();
+
+    auto upgradeSetKey = loadGen.getConfigUpgradeSetKey(
+        createUpgradeLoadGenConfig.getSorobanUpgradeConfig(), contractId);
 
     numTxsBefore = getSuccessfulTxCount();
     loadGen.generateLoad(createUpgradeLoadGenConfig);
